# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StreamChunkRequest": {
        "description": "v4.0 — 5초 PCM chunk 업로드 (draft).\n\n실제 구현은 multipart/form-data 또는 WebSocket binary frame 로 전환 가능.",
        "properties": {
          "audio_base64": {
            "description": "16kHz mono PCM 또는 wav, base64",
            "title": "Audio Base64",
            "type": "string"
          },
          "chunk_index": {
            "minimum": 0.0,
            "title": "Chunk Index",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "chunk_index",
          "audio_base64"
        ],
        "title": "StreamChunkRequest",
        "type": "object"
      },
      "StreamChunkResponse": {
        "description": "chunk 검출 즉답 — 의도 분류 + 누적 신호 개수 (draft).\n\nIdentity: 점수·등급 X. 검출된 메타인식·민감정보·송금동의 신호 누적 개수만.",
        "properties": {
          "alert": {
            "description": "임계 초과 여부 — true 면 클라이언트가 경보 트리거",
            "title": "Alert",
            "type": "boolean"
          },
          "alert_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "경보 사유 (alert=true 일 때만)",
            "title": "Alert Reason"
          },
          "chunk_index": {
            "title": "Chunk Index",
            "type": "integer"
          },
          "cumulative_signal_count": {
            "description": "현재까지 누적된 위험 신호 총 개수 (v4.1 슬라이딩 윈도우)",
            "minimum": 0.0,
            "title": "Cumulative Signal Count",
            "type": "integer"
          },
          "label": {
            "description": "experiments/v4_intent Haiku 분류기",
            "enum": [
              "META_AWARE",
              "SENSITIVE_INFO",
              "TRANSFER_AGREE",
              "NORMAL"
            ],
            "title": "Label",
            "type": "string"
          },
          "transcript": {
            "description": "5초 chunk Whisper 결과",
            "title": "Transcript",
            "type": "string"
          }
        },
        "required": [
          "chunk_index",
          "transcript",
          "label",
          "cumulative_signal_count",
          "alert"
        ],
        "title": "StreamChunkResponse",
        "type": "object"
      },
      "StreamEndResponse": {
        "description": "통화 종료 후 사후 검출 결과 (draft).\n\nPhase 4 (Serper 검증) 까지 수행한 DetectionReport 와 동등.\nIdentity: 점수·등급 X — detected_signals[] 만 보고. 통합 기업이 자체 판정.",
        "properties": {
          "alert_count": {
            "description": "통화 도중 발동된 경보 횟수",
            "title": "Alert Count",
            "type": "integer"
          },
          "analysis_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "DB 저장된 경우 — `/api/result/{token}` 으로 카드 조회 가능",
            "title": "Analysis Run Id"
          },
          "detected_signals": {
            "description": "시간순 검출 신호 list (각 신호의 학술/법적 근거 포함)",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Detected Signals",
            "type": "array"
          },
          "disclaimer": {
            "description": "ScamGuardian 은 사기 판정을 내리지 않습니다 안내",
            "title": "Disclaimer",
            "type": "string"
          },
          "full_transcript": {
            "title": "Full Transcript",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "summary": {
            "description": "'위험 신호 N개 검출되었습니다.' 류 요약",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "full_transcript",
          "detected_signals",
          "alert_count",
          "summary",
          "disclaimer"
        ],
        "title": "StreamEndResponse",
        "type": "object"
      },
      "StreamSessionDetail": {
        "description": "세션 상태 조회 응답 (draft). 점수·등급 표면 X.",
        "properties": {
          "chunks_received": {
            "title": "Chunks Received",
            "type": "integer"
          },
          "cumulative_signal_count": {
            "description": "현재까지 누적된 위험 신호 총 개수",
            "minimum": 0.0,
            "title": "Cumulative Signal Count",
            "type": "integer"
          },
          "last_chunk_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Chunk At"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "number"
          },
          "status": {
            "enum": [
              "active",
              "ended",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          },
          "transcript_so_far": {
            "title": "Transcript So Far",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "chunks_received",
          "cumulative_signal_count",
          "started_at",
          "last_chunk_at",
          "transcript_so_far"
        ],
        "title": "StreamSessionDetail",
        "type": "object"
      },
      "StreamStartRequest": {
        "description": "v4.0 — 통화 세션 생성 요청 (draft).",
        "properties": {
          "chunk_sec": {
            "default": 5,
            "description": "Whisper chunk 길이. v4.0 권장 5s",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Chunk Sec",
            "type": "integer"
          },
          "locale": {
            "default": "ko",
            "description": "STT 언어 코드 (기본 ko)",
            "title": "Locale",
            "type": "string"
          },
          "user_id": {
            "description": "카카오 user.id 또는 외부 식별자",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "StreamStartRequest",
        "type": "object"
      },
      "StreamStartResponse": {
        "properties": {
          "expires_at": {
            "description": "epoch sec — 세션 자동 종료 시각",
            "title": "Expires At",
            "type": "number"
          },
          "issued_at": {
            "description": "epoch sec",
            "title": "Issued At",
            "type": "number"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "upload_url": {
            "description": "chunk POST 대상 (보통 같은 호스트)",
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "issued_at",
          "expires_at",
          "upload_url"
        ],
        "title": "StreamStartResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "**한국어 사기 신호 검출 reference implementation** — 음성·텍스트·이미지·PDF·URL 입력에서 위험 신호를 검출하고 각 신호의 학술/법적 근거를 transparent 하게 보고한다.\n\n⚠️ **Identity Boundary** (CLAUDE.md): ScamGuardian 은 *사기 판정을 내리지 않는다*. VirusTotal 이 70개 백신의 검출 결과를 보고만 하는 모델과 동일. 판정 logic 은 통합한 기업(통신사·은행·메신저 앱)이 자기 risk tolerance 에 따라 구현한다.\n\n- **Public 4 endpoint** (`Public` 태그) — 외부 통합 시작점, `DetectionReport` 응답\n- **Webhook** (`Webhook` 태그) — 카카오 오픈빌더\n- **Admin** (3 태그) — 라벨링 / platform / training\n- **v4 draft** — Live Call Guard 설계 preview (구현 X)\n\n통합 가이드: [`docs/INTEGRATION_GUIDE.md`](https://github.com/example/scamguardian-v2/blob/main/docs/INTEGRATION_GUIDE.md)",
    "title": "ScamGuardian Signal Detection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/result/{token}": {
      "get": {
        "description": "카카오 챗봇 응답 카드의 '자세히 보기' 링크가 호출하는 엔드포인트. 결과 발급 시점에 1시간 유효 토큰을 발급(`secrets.token_urlsafe(16)`)하고 메모리에 저장한다. 인메모리 저장이라 서버 재시작 시 토큰은 모두 만료.\n\n**응답**:\n- `result` — `DetectionReport.to_dict()` 전체 (`/api/analyze` 와 동일 schema)\n- `user_context` — 사용자가 챗봇과 나눈 Q&A 요약 (있을 때)\n- `input_type` — `TEXT|URL|VIDEO|FILE|IMAGE|PDF`\n- `chat_history` — 봇/사용자 대화 시간순 전체\n- `flag_rationale` — 검출된 신호별 `{rationale, source}` 매핑\n- `expires_at` — 토큰 만료 시각 (epoch sec)\n\n**인증**: 선택. 토큰 자체가 인증 역할 (URL 알면 누구나 조회).\n\n**에러**:\n- `404` — 토큰이 존재한 적 없음 또는 서버 재시작\n- `410` — 토큰 만료 (1시간 경과)\n\n**curl**:\n```bash\ncurl https://api.example.com/api/result/abcd1234efgh5678\n```",
        "operationId": "get_result_by_token_api_result__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Result By Token Api Result  Token  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "토큰 없음"
          },
          "410": {
            "description": "토큰 만료 — 결과 재요청 필요"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "결과 토큰으로 분석 결과 조회",
        "tags": [
          "Public"
        ]
      }
    },
    "/api/v4/stream/chunk": {
      "post": {
        "description": "한 chunk 단위로 STT (`whisper-1` 또는 로컬) + Haiku 의도 분류 (`META_AWARE | SENSITIVE_INFO | TRANSFER_AGREE | NORMAL`) 수행.\n\n임계 초과 시 응답에 `alert=true` + `alert_reason` 포함 → 클라이언트가 경보음·진동 트리거. 구현 시 WebSocket push 로 양방향 채널 권장.\n\n⚠️ **v4 draft — design preview only, not yet implemented.** 호출 시 `501 Not Implemented` 반환. schema 검토 + 외부 클라이언트 통합 설계용.",
        "operationId": "stream_chunk_api_v4_stream_chunk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamChunkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamChunkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "v4 미구현 — design preview"
          }
        },
        "summary": "[draft] 5초 chunk 업로드 + 즉시 의도 분류",
        "tags": [
          "v4 (draft)"
        ]
      }
    },
    "/api/v4/stream/end": {
      "post": {
        "description": "사용자가 통화 종료 → 누적 transcript 로 Phase 4 (Serper 교차 검증) + Phase 5 (스코어링) 수행. 결과는 `analysis_run_id` 로 DB 저장되어 `/api/result/{token}` 카드 발급 가능.\n\n⚠️ **v4 draft — design preview only, not yet implemented.** 호출 시 `501 Not Implemented` 반환. schema 검토 + 외부 클라이언트 통합 설계용.",
        "operationId": "stream_end_api_v4_stream_end_post",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamEndResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "v4 미구현 — design preview"
          }
        },
        "summary": "[draft] 통화 종료 + 사후 분석",
        "tags": [
          "v4 (draft)"
        ]
      }
    },
    "/api/v4/stream/start": {
      "post": {
        "description": "통화 시작 시점에 호출 — session_id 발급 + chunk 업로드 endpoint 안내.\n\n구현 시 웹앱이 `getUserMedia` → AudioWorklet 으로 16kHz mono PCM 5초 buffer 생성 후 `/api/v4/stream/chunk` 로 순차 전송.\n\n⚠️ **v4 draft — design preview only, not yet implemented.** 호출 시 `501 Not Implemented` 반환. schema 검토 + 외부 클라이언트 통합 설계용.",
        "operationId": "stream_start_api_v4_stream_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "v4 미구현 — design preview"
          }
        },
        "summary": "[draft] Live Call 세션 시작",
        "tags": [
          "v4 (draft)"
        ]
      }
    },
    "/api/v4/stream/{session_id}": {
      "get": {
        "description": "현재 누적 transcript / 누적 위험도 / chunk 개수. 폴링 또는 모니터링 UI 용.\n\n⚠️ **v4 draft — design preview only, not yet implemented.** 호출 시 `501 Not Implemented` 반환. schema 검토 + 외부 클라이언트 통합 설계용.",
        "operationId": "stream_get_api_v4_stream__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamSessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "v4 미구현 — design preview"
          }
        },
        "summary": "[draft] 세션 상태 조회",
        "tags": [
          "v4 (draft)"
        ]
      }
    },
...
  },
...
}